backend/requirements.txt
asgiref==3.7.2
distlib==0.3.8
Django==5.0.3
filelock==3.13.1
platformdirs==4.2.0
setuptools==69.2.0
sqlparse==0.4.4
virtualenv==20.25.1
wheel==0.43.0

djangorestframework~=3.15.0
pillow~=10.2.0
channels~=4.1.0
sympy~=1.12